requirements.txt
asgiref==3.8.1
bootstrap4==0.1.0
contourpy==1.2.1
crispy-bootstrap5==2024.2
cycler==0.12.1
Django==5.0.6
django-crispy-forms==2.2
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1
mpmath==1.3.0
networkx==3.3
numpy==2.0.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pillow==10.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
sympy==1.12.1
torch==2.3.1
torchvision==0.18.1
triton==2.3.1
typing_extensions==4.12.2
